+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Patch dnsConfig and dnsPolicy of chart-operator-unique deployment.
+
 ## [0.4.0] - 2020-10-30
 
 ### Added